openbb_platform/extensions/quantitative/pyproject.toml
[tool.poetry]
name = "openbb-quantitative"
version = "1.3.0"
description = "Quantitative Analysis extension for OpenBB"
authors = ["OpenBB Team <hello@openbb.co>"]
license = "AGPL-3.0-only"
readme = "README.md"
packages = [{ include = "openbb_quantitative" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"  # scipy forces python <4.0 explicitly
scipy = "^1.10.1"
statsmodels = "^0.14.0"
pandas-ta = "^0.3.14b"
openbb-core = "^1.3.0"
numpy = "^1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."openbb_core_extension"]
quantitative = "openbb_quantitative.quantitative_router:router"
